requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
Werkzeug==3.0.1
python-docx==1.2.0
pyTelegramBotAPI==4.14.0
psycopg2-binary==2.9.9

